test: move example data generation into tests
Instead of creating one-size-fits all content for the test ES DB,
have each test create the data it needs.
