+++ src/test/java/de/komoot/photon/ApiIntegrationTest.java
@@ -1,8 +1,10 @@
 package de.komoot.photon;
 
+import com.google.common.collect.ImmutableMap;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 import java.io.BufferedReader;
@@ -15,11 +17,24 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static spark.Spark.*;
 
+import de.komoot.photon.PhotonDoc;
+import de.komoot.photon.elasticsearch.Importer;
+
 /**
  * These test connect photon to an already running ES node (setup in ESBaseTester) so that we can directly test the API
  */
 public class ApiIntegrationTest extends ESBaseTester {
     private static final int LISTEN_PORT = 30234;
+
+    @Before
+    public void setUp() throws Exception {
+        setUpES();
+        Importer instance = new Importer(getClient(), "en");
+        instance.add(createDoc(13.38886, 52.51704, 1000, 1000, "place", "city"));
+        instance.add(createDoc(13.39026, 52.54714, 1001, 1001, "place", "town"));
+        instance.finish();
+        refresh();
+    }
 
     @After
     public void shutdown() {
@@ -33,7 +48,7 @@ public class ApiIntegrationTest extends ESBaseTester {
      */
     @Test
     public void testNoCors() throws Exception {
-        App.main(new String[]{"-cluster", clusterName, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1"});
+        App.main(new String[]{"-cluster", TEST_CLUSTER_NAME, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1"});
         awaitInitialization();
         HttpURLConnection connection = (HttpURLConnection) new URL("http://127.0.0.1:" + port() + "/api?q=berlin").openConnection();
         assertNull(connection.getHeaderField("Access-Control-Allow-Origin"));
@@ -44,7 +59,7 @@ public class ApiIntegrationTest extends ESBaseTester {
      */
     @Test
     public void testCorsAny() throws Exception {
-        App.main(new String[]{"-cluster", clusterName, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1",
+        App.main(new String[]{"-cluster", TEST_CLUSTER_NAME, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1",
                 "-cors-any"});
         awaitInitialization();
         HttpURLConnection connection = (HttpURLConnection) new URL("http://127.0.0.1:" + port() + "/api?q=berlin").openConnection();
@@ -56,7 +71,7 @@ public class ApiIntegrationTest extends ESBaseTester {
      */
     @Test
     public void testCorsOriginIsSetToSpecificDomain() throws Exception {
-        App.main(new String[]{"-cluster", clusterName, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1",
+        App.main(new String[]{"-cluster", TEST_CLUSTER_NAME, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1",
                 "-cors-origin", "www.poole.ch"});
         awaitInitialization();
         HttpURLConnection connection = (HttpURLConnection) new URL("http://127.0.0.1:" + port() + "/api?q=berlin").openConnection();
@@ -65,7 +80,7 @@ public class ApiIntegrationTest extends ESBaseTester {
 
     @Test
     public void testSearchForBerlin() throws Exception {
-        App.main(new String[]{"-cluster", clusterName, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1"});
+        App.main(new String[]{"-cluster", TEST_CLUSTER_NAME, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1"});
         awaitInitialization();
         HttpURLConnection connection = (HttpURLConnection) new URL("http://127.0.0.1:" + port() + "/api?q=berlin&limit=1").openConnection();
         JSONObject json = new JSONObject(
@@ -75,8 +90,8 @@ public class ApiIntegrationTest extends ESBaseTester {
         JSONObject feature = features.getJSONObject(0);
         JSONObject properties = feature.getJSONObject("properties");
         assertEquals("way", properties.getString("osm_type"));
-        assertEquals("tourism", properties.getString("osm_key"));
-        assertEquals("attraction", properties.getString("osm_value"));
+        assertEquals("place", properties.getString("osm_key"));
+        assertEquals("city", properties.getString("osm_value"));
         assertEquals("berlin", properties.getString("name"));
     }
 
@@ -85,7 +100,7 @@ public class ApiIntegrationTest extends ESBaseTester {
      */
     @Test
     public void testApiWithLocationBias() throws Exception {
-        App.main(new String[]{"-cluster", clusterName, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1"});
+        App.main(new String[]{"-cluster", TEST_CLUSTER_NAME, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1"});
         awaitInitialization();
         HttpURLConnection connection = (HttpURLConnection) new URL("http://127.0.0.1:" + port() + "/api?q=berlin&limit=1&lat=52.54714&lon=13.39026")
                 .openConnection();
@@ -96,8 +111,8 @@ public class ApiIntegrationTest extends ESBaseTester {
         JSONObject feature = features.getJSONObject(0);
         JSONObject properties = feature.getJSONObject("properties");
         assertEquals("way", properties.getString("osm_type"));
-        assertEquals("railway", properties.getString("osm_key"));
-        assertEquals("station", properties.getString("osm_value"));
+        assertEquals("place", properties.getString("osm_key"));
+        assertEquals("town", properties.getString("osm_value"));
         assertEquals("berlin", properties.getString("name"));
     }
 
@@ -106,7 +121,7 @@ public class ApiIntegrationTest extends ESBaseTester {
      */
     @Test
     public void testApiReverse() throws Exception {
-        App.main(new String[]{"-cluster", clusterName, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1"});
+        App.main(new String[]{"-cluster", TEST_CLUSTER_NAME, "-listen-port", Integer.toString(LISTEN_PORT), "-transport-addresses", "127.0.0.1"});
         awaitInitialization();
         HttpURLConnection connection = (HttpURLConnection) new URL("http://127.0.0.1:" + port() + "/reverse/?lon=13.38886&lat=52.51704").openConnection();
         JSONObject json = new JSONObject(
@@ -116,8 +131,8 @@ public class ApiIntegrationTest extends ESBaseTester {
         JSONObject feature = features.getJSONObject(0);
         JSONObject properties = feature.getJSONObject("properties");
         assertEquals("way", properties.getString("osm_type"));
-        assertEquals("tourism", properties.getString("osm_key"));
-        assertEquals("attraction", properties.getString("osm_value"));
+        assertEquals("place", properties.getString("osm_key"));
+        assertEquals("city", properties.getString("osm_value"));
         assertEquals("berlin", properties.getString("name"));
     }
 }
