+++ src/test/java/de/komoot/photon/query/PhotonQueryBuilderSearchTest.java
@@ -5,10 +5,18 @@ import static org.junit.Assert.assertEquals;
 import java.io.IOException;
 import java.util.Set;
 
+import com.google.common.collect.ImmutableList;
+import com.vividsolutions.jts.geom.Coordinate;
+import com.vividsolutions.jts.geom.GeometryFactory;
+import com.vividsolutions.jts.geom.Point;
+import com.vividsolutions.jts.geom.PrecisionModel;
+import de.komoot.photon.PhotonDoc;
+import de.komoot.photon.elasticsearch.Importer;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.action.search.SearchType;
 import org.elasticsearch.client.Client;
 import org.elasticsearch.index.query.QueryBuilder;
+import org.junit.Before;
 import org.junit.Test;
 
 import com.google.common.collect.ImmutableMap;
@@ -20,6 +28,31 @@ import de.komoot.photon.ESBaseTester;
  * Created by Sachin Dole on 2/20/2015.
  */
 public class PhotonQueryBuilderSearchTest extends ESBaseTester {
+
+    @Before
+    public void setUp() throws Exception {
+        setUpES();
+        ImmutableList<String> tags = ImmutableList.of("tourism", "attraction", "tourism", "hotel", "tourism", "museum", "tourism", "information", "amenity",
+                "parking", "amenity", "restaurant", "amenity", "information", "food", "information", "railway", "station");
+        Importer instance = new Importer(getClient(), "en");
+        double lon = 13.38886;
+        double lat = 52.51704;
+        for (int i = 0; i < tags.size(); i++) {
+            String key = tags.get(i);
+            String value = tags.get(++i);
+            PhotonDoc doc = this.createDoc(lon, lat, i, i, key, value);
+            instance.add(doc);
+            lon += 0.00004;
+            lat += 0.00086;
+            doc = this.createDoc(lon, lat, i + 1, i + 1, key, value);
+            instance.add(doc);
+            lon += 0.00004;
+            lat += 0.00086;
+        }
+        instance.finish();
+        refresh();
+    }
+
 
     /**
      * Find me all places named "berlin" that are tagged "tourism=attraction"
